+++ src/_features/Product/components/index.ts
@@ -1,0 +1,2 @@
+export { Product } from "./Product";
+export { ProductError } from "./ProductError";
