+++ src/_features/ShoppingStore/index.ts
@@ -1,0 +1,2 @@
+export { ShoppingCartContext } from "./context/ShoppingCartContext";
+export { Store } from "./page/StorePage";
